requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.2.8
django-debug-toolbar==3.2.2
django-elasticsearch-dsl==7.2.0
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
elasticsearch==7.15.1
elasticsearch-dsl==7.4.0
flatbuffers==2.0
idna==3.3
imageio==2.9.0
inflection==0.5.1
instance==0.0.0.0.0.0a0.dev20210120
itypes==1.2.0
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
onnxruntime==1.9.0
opencv-python-headless==4.5.3.56
packaging==21.0
pandas==1.3.4
Pillow==8.2.0
protobuf==3.18.1
psycopg2-binary==2.9.1
pydantic==1.7.4
pyparsing==2.4.7
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-image==0.18.3
scipy==1.7.0
six==1.16.0
sqlparse==0.4.2
tifffile==2021.10.12
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
